Remove the mm table TCA
This patch (https://github.com/TYPO3/TYPO3.CMS/commit/65a9d8a7a423979777d2c4634565ae690dc477be)
removed a core check for the "pid" field on a table. The core now only checks for the
TCA. If it finds TCA it will add the pid clause.

I think we don't need the TCA at all and it is only an autoloader artifact.